Refactor tooltip setup as component
Added property to decide whether tooltip to display
features or not. This will be required if tooltip is
used to draw shape.
